feat(csharp/src/Drivers/Apache): improve type name handling for CHAR/VARCHAR/DECIMAL.
